Add a depends_on:PACKAGE criterion
pub_crawl seems like a useful tool to get a rough idea about the
amount of fallout from a breaking change to a package.  It therefore
would be useful to allow it to find dependent packages for packages
other than flutter.

Generalize the `flutter` criterion to a new `depends_on:PACKAGE`
criterion, and make the `flutter` one to leverage it.

(Note that this is very slow since it iterates over all packages in
pub.dev.  A better long-term approach probably would be to submit a
`dependency:PACKAGE` query to pub.dev and to process the results.)

Additionally make pub_crawl fail if given unrecognized criteria.
(This isn't as friendly as it should be, but it's better than
proceeding on an unwanted and expensive operation.)